requirements.dev.txt
autopep8
black
boto3
dj-database-url
Django==4.0
django-cors-headers
django-heroku  
django-summernote
django-utils-six
djangorestframework
djangorestframework-simplejwt
django-rest-authemail
django-storages
flake8
gunicorn
pylint
psycopg2
psycopg2-binary
pytest
pytest-django
python-dotenv
requests
six
whitenoise